// services/onenote-mcp/src/unique/unique.dtos.ts
const CreateScopeByPathsSchema = z.object({
  paths: z.array(z.string()),
  inheritAccess: z.boolean().default(true),
});

const CreateScopeByParentSchema = z.object({
  parentScopeId: z.string(),
  relativePaths: z.array(z.string()),
  inheritAccess: z.boolean().default(true),
});

export const PublicCreateScopeRequestSchema = z.union([
  CreateScopeByPathsSchema,
  CreateScopeByParentSchema,
]);
export type PublicCreateScopeRequest = z.infer<typeof PublicCreateScopeRequestSchema>;

export const ScopeSchema = z.object({
  id: z.string(),
  name: z.string(),
  parentId: z.string().nullable().optional(),
  object: z.literal('folder'),
});
export type Scope = z.infer<typeof ScopeSchema>;

export const PublicCreateScopeResultSchema = z.object({
  createdFolders: z.array(ScopeSchema).default([]),
});
export type PublicCreateScopeResult = z.infer<typeof PublicCreateScopeResultSchema>;

// !SECTION - CreateScope endpoint types

// SECTION - AddScopeAccess endpoint types

export enum ScopeAccessType {
  Manage = 'MANAGE',
  Read = 'READ',
  Write = 'WRITE',
}

export enum ScopeAccessEntityType {
  Group = 'GROUP',
  User = 'USER',
}

export const PublicScopeAccessRequestSchema = z.object({
  entityId: z.string(),
  entityType: z.enum(ScopeAccessEntityType),
  type: z.enum(ScopeAccessType),
});
export type PublicScopeAccessSchema = z.infer<typeof PublicScopeAccessRequestSchema>;
export type PublicScopeAccessRequest = z.infer<typeof PublicScopeAccessRequestSchema>;

export const PublicAddScopeAccessRequestSchema = z
  .object({
    scopeId: z.string().optional(),
    scopePath: z.string().optional(),
    scopeAccesses: z.array(PublicScopeAccessRequestSchema),
    applyToSubScopes: z.boolean().default(false),
  })
  .refine((data) => data.scopeId || data.scopePath, {
    message: 'scopeId or scopePath must be provided.',
  });
export type PublicAddScopeAccessRequest = z.infer<typeof PublicAddScopeAccessRequestSchema>;

export const ChildrenScopeSchema = z.object({
  id: z.string(),
  name: z.string(),
});
export type ChildrenScope = z.infer<typeof ChildrenScopeSchema>;

export const PublicAddScopeAccessResultSchema = z.object({
  id: z.string(),
  name: z.string(),
  scopeAccesses: z.array(PublicScopeAccessRequestSchema),
  children: z.array(ChildrenScopeSchema),
  object: z.literal('updateFolderAccessResult'),
});
export type PublicAddScopeAccessResult = z.infer<typeof PublicAddScopeAccessResultSchema>;